% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_rapid.R
\name{GWASapply_rapid}
\alias{GWASapply_rapid}
\title{Rapid GWAS with PCA}
\usage{
GWASapply_rapid(
  pheno = NULL,
  geno = NULL,
  Cov = NULL,
  PCA.M = 3,
  cutoff = NULL
)
}
\arguments{
\item{pheno}{file with numeric phenotypic values}

\item{geno}{data.frame with genotype calls coded as 0,1,2.}

\item{Cov}{numeric data.frame with covariates values}

\item{PCA.M}{number of principal components to use default is 3}

\item{cutoff}{If cutoff is default, uses Bonferroni; 0.05/number of SNPs}
}
\value{
GWAS P-values
}
\description{
Rapid GWAS with PCA
}
